Implemented conversion from f16 to f8e4m3

+++ src/plugins/intel_cpu/tests/functional/shared_tests_instances/single_layer_tests/conversion.cpp
@@ -18,7 +18,8 @@ const std::vector<ov::test::utils::ConversionTypes> conversionOpTypes = {
 const std::vector<std::vector<ov::Shape>> shapes = {{{1, 2, 3, 4}}};
 
 const std::vector<ov::element::Type> inTypes = {
-    ov::element::f32,
+    // ov::element::f32,
+    ov::element::f16,
 };
 
 const std::vector<ov::element::Type> outTypes = {
